# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DiseaseX Triage & Protocol Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PatientStatus = (typeof PatientStatus)[keyof typeof PatientStatus];

export const PatientStatus = {
  active: "active",
  discharged: "discharged",
  referred: "referred",
} as const;

export interface Patient {
  id: number;
  name: string;
  /** @nullable */
  ageYears?: number | null;
  /** @nullable */
  ageMonths?: number | null;
  /** @nullable */
  weightKg?: number | null;
  location: string;
  status: PatientStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePatientBody {
  name: string;
  /** @nullable */
  ageYears?: number | null;
  /** @nullable */
  ageMonths?: number | null;
  /** @nullable */
  weightKg?: number | null;
  location: string;
}

export type UpdatePatientBodyStatus =
  (typeof UpdatePatientBodyStatus)[keyof typeof UpdatePatientBodyStatus];

export const UpdatePatientBodyStatus = {
  active: "active",
  discharged: "discharged",
  referred: "referred",
} as const;

export interface UpdatePatientBody {
  status?: UpdatePatientBodyStatus;
  /** @nullable */
  weightKg?: number | null;
}

export interface PatternMatch {
  disease: string;
  confidencePercent: number;
  reasoning: string;
  /** @nullable */
  protocolId?: number | null;
}

export interface Assessment {
  id: number;
  patientId: number;
  symptoms: string[];
  /** @nullable */
  tempCelsius?: number | null;
  /** @nullable */
  heartRateBpm?: number | null;
  /** @nullable */
  spo2Percent?: number | null;
  /** @nullable */
  durationDays?: number | null;
  patternMatches: PatternMatch[];
  isAtypical: boolean;
  /** @nullable */
  firstDoseAt?: string | null;
  /** @nullable */
  firstDoseDrug?: string | null;
  /** @nullable */
  referralAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateAssessmentBody {
  patientId: number;
  symptoms: string[];
  /** @nullable */
  tempCelsius?: number | null;
  /** @nullable */
  heartRateBpm?: number | null;
  /** @nullable */
  spo2Percent?: number | null;
  /** @nullable */
  durationDays?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface LogFirstDoseBody {
  drug: string;
  /** @nullable */
  notes?: string | null;
}

export type ProtocolSeverity =
  (typeof ProtocolSeverity)[keyof typeof ProtocolSeverity];

export const ProtocolSeverity = {
  mild: "mild",
  moderate: "moderate",
  severe: "severe",
  critical: "critical",
} as const;

export interface Protocol {
  id: number;
  disease: string;
  title: string;
  severity: ProtocolSeverity;
  immediateAction: string;
  steps: string[];
  referralThreshold: string;
  /** @nullable */
  dosageFormula?: string | null;
  drugs: string[];
  whoBased: boolean;
  createdAt: string;
}

export type OutbreakAlertStatus =
  (typeof OutbreakAlertStatus)[keyof typeof OutbreakAlertStatus];

export const OutbreakAlertStatus = {
  active: "active",
  resolved: "resolved",
  monitoring: "monitoring",
} as const;

export interface OutbreakAlert {
  id: number;
  assessmentId: number;
  location: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  symptomsSummary: string;
  status: OutbreakAlertStatus;
  reportedAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}

export interface CreateOutbreakAlertBody {
  assessmentId: number;
  location: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  symptomsSummary: string;
}

export interface DashboardSummary {
  activeCases: number;
  assessmentsToday: number;
  activeOutbreakAlerts: number;
  criticalCases: number;
  referralsPending: number;
  firstDosesAdministered: number;
  atypicalCasesThisWeek: number;
}

export type PatientQueueEntryUrgency =
  (typeof PatientQueueEntryUrgency)[keyof typeof PatientQueueEntryUrgency];

export const PatientQueueEntryUrgency = {
  low: "low",
  monitor: "monitor",
  critical: "critical",
} as const;

export interface PatientQueueEntry {
  patient: Patient;
  latestAssessment: Assessment;
  topMatch?: PatternMatch;
  urgency: PatientQueueEntryUrgency;
}

export type ListPatientsParams = {
  status?: ListPatientsStatus;
};

export type ListPatientsStatus =
  (typeof ListPatientsStatus)[keyof typeof ListPatientsStatus];

export const ListPatientsStatus = {
  active: "active",
  discharged: "discharged",
  referred: "referred",
  all: "all",
} as const;

export type ListAssessmentsParams = {
  patientId?: number;
};

export type ListProtocolsParams = {
  disease?: string;
  search?: string;
};