# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DiseaseX Triage & Protocol Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  activeCases: number;
  assessmentsToday: number;
  activeOutbreakAlerts: number;
  criticalCases: number;
  referralsPending: number;
  firstDosesAdministered: number;
  atypicalCasesThisWeek: number;
}