# lib/api-zod/src/generated/types/logFirstDoseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DiseaseX Triage & Protocol Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface LogFirstDoseBody {
  drug: string;
  /** @nullable */
  notes?: string | null;
}