# lib/api-zod/src/generated/types/outbreakAlert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DiseaseX Triage & Protocol Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { OutbreakAlertStatus } from "./outbreakAlertStatus";

export interface OutbreakAlert {
  id: number;
  assessmentId: number;
  location: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  symptomsSummary: string;
  status: OutbreakAlertStatus;
  reportedAt: Date;
  /** @nullable */
  resolvedAt?: Date | null;
}