# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DiseaseX Triage & Protocol Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientStatus } from "./patientStatus";

export interface Patient {
  id: number;
  name: string;
  /** @nullable */
  ageYears?: number | null;
  /** @nullable */
  ageMonths?: number | null;
  /** @nullable */
  weightKg?: number | null;
  location: string;
  status: PatientStatus;
  createdAt: Date;
  updatedAt: Date;
}